requirements.txt
uvicorn==0.13.3
rootpath==0.1.1
SQLAlchemy==1.3.20
FastAPI_SQLAlchemy==0.2.1
fastapi_login==1.5.2
pydantic==1.7.3
fastapi==0.63.0
python-dotenv==0.15.0
